+++ dart/lib/vm_service_lib.dart
@@ -42,6 +42,22 @@ String _printEnum(Object obj) {
   return str.substring(index + 1);
 }
 
+List<BoundVariable> _parseBoundVariables(json) {
+  if (json is List) {
+    return json.map(BoundVariable._parse).toList();
+  } else {
+    return [];
+  }
+}
+
+List<BoundField> _parseBoundFields(json) {
+  if (json is List) {
+    return json.map(BoundField._parse).toList();
+  } else {
+    return [];
+  }
+}
+
 Map<String, Function> _typeFactories = {
   'BoundField': BoundField._parse,
   'BoundVariable': BoundVariable._parse,
@@ -1420,7 +1436,7 @@ class Frame extends Response {
     function = _createObject(json['function']);
     code = _createObject(json['code']);
     location = _createObject(json['location']);
-    vars = _createObject(json['vars']);
+    vars = _parseBoundVariables(json['vars']);
   }
 
   int index;
@@ -1619,7 +1635,7 @@ class Instance extends Obj {
     name = json['name'];
     typeClass = _createObject(json['typeClass']);
     parameterizedClass = _createObject(json['parameterizedClass']);
-    fields = _createObject(json['fields']);
+    fields = _parseBoundFields(json['fields']);
     elements = _createObject(json['elements']);
     associations = _createObject(json['associations']);
     bytes = json['bytes'];