+++ dart/tool/dart/generate_dart.dart
@@ -266,6 +266,22 @@ String _printEnum(Object obj) {
   return str.substring(index + 1);
 }
 
+List<BoundVariable> _parseBoundVariables(json) {
+  if (json is List) {
+    return json.map(BoundVariable._parse).toList();
+  } else {
+    return [];
+  }
+}
+
+List<BoundField> _parseBoundFields(json) {
+  if (json is List) {
+    return json.map(BoundField._parse).toList();
+  } else {
+    return [];
+  }
+}
+
 ''');
     gen.writeln();
     gen.write('Map<String, Function> _typeFactories = {');
@@ -560,6 +576,12 @@ class Type extends Member {
         String enumTypeName = field.type.types.first.name;
         gen.writeln(
           "${field.generatableName} = _parse${enumTypeName}[json['${field.name}']];");
+      } else if (name == 'Frame' && field.name == 'vars') {
+        // TODO: Reported as https://github.com/dart-lang/sdk/issues/24654.
+        gen.writeln("${field.generatableName} = _parseBoundVariables(json['${field.name}']);");
+      } else if (name == 'Instance' && field.name == 'fields') {
+        // TODO: Reported as https://github.com/dart-lang/sdk/issues/24654.
+        gen.writeln("${field.generatableName} = _parseBoundFields(json['${field.name}']);");
       } else {
         gen.writeln("${field.generatableName} = _createObject(json['${field.name}']);");
       }